+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- tests: fixed `TypeError` when converting to `pathlib.Path` on python3.5
+  by replacing `tmpdir` fixture with `temp_path`
 
 ## [0.1.1] - 2020-02-06
 ### Fixed